moved treeMock file
